pyproject.toml
[project]
name = "warnet-bot"
version = "1.11.0"
description = "A bot for Warga Nusantara Teyvat (WARNET) discord server"
requires-python = ">=3.11,<3.12"
dependencies = [
    "anyio>=4.9.0",
    "asyncpg==0.27.0",
    "discord-py==2.5.2",
    "hoyolab-rss-feeds==2.4.2",
    "imagetext-py==2.1.3",
    "pillow==10.0.0",
    "python-dotenv==0.21.1",
    "pytz==2022.7.1",
    "ratelimit==2.2.1",
    "taskipy==1.12.0",
]

[dependency-groups]
dev = ["pytest==7.4.1", "pytest-cov==4.1.0", "ruff>=0.12.4", "pyright>=1.1.403"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # ignore all pydocstyle
    "D",

    "T201", # enable print
    "PLR0911", # allow many return
    "ERA001", # allow commented out code
    "BLE001", # allow blind except

    # conflict with formatter https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "E501",
]

[tool.pyright]
include = ["bot"]
ignore = ["tests"]

[tool.taskipy.tasks]
start = "python -m bot"
